feat: Add hide-delete and hide-avatar parameters (#35)
* feat: Add hide delete card parameter

* feat: Add hide avatar parameter to SkeletonLoader

* fix: Fix avatar padding issue

+++ siren-sdk/src/main/java/com/keyvalue/siren/androidsdk/core/elements/NotificationCard.kt
@@ -126,10 +126,6 @@ fun NotificationCard(
                             .weight(1f)
                             .semantics { contentDescription = "siren-notification-avatar-${notification?.id}" },
                 )
-            } else {
-                Spacer(
-                    modifier = Modifier.size(notificationCardStyle.avatarSize!!).weight(1f),
-                )
             }
             Column(
                 modifier =
@@ -201,17 +197,19 @@ fun NotificationCard(
                 }
             }
 
-            Box(modifier = Modifier.weight(1f).padding(end = 6.dp), contentAlignment = Alignment.CenterEnd) {
-                Icon(
-                    imageVector = Icons.Default.Clear,
-                    contentDescription = "siren-notification-delete-${notification?.id}",
-                    tint = themeColors?.deleteIcon!!,
-                    modifier =
-                        Modifier
-                            .size(notificationCardStyle.deleteIconSize!!)
-                            .clickable { deleteNotificationCallback() }
-                            .semantics { contentDescription = "siren-notification-delete-${notification?.id}" },
-                )
+            if (cardProps?.hideDelete != true) {
+                Box(modifier = Modifier.weight(1f).padding(end = 6.dp), contentAlignment = Alignment.CenterEnd) {
+                    Icon(
+                        imageVector = Icons.Default.Clear,
+                        contentDescription = "siren-notification-delete-${notification?.id}",
+                        tint = themeColors?.deleteIcon!!,
+                        modifier =
+                            Modifier
+                                .size(notificationCardStyle.deleteIconSize!!)
+                                .clickable { deleteNotificationCallback() }
+                                .semantics { contentDescription = "siren-notification-delete-${notification?.id}" },
+                    )
+                }
             }
         }
     }
